➕ Move calendar event classes into corresponding components
