+++ frontend/src/elements/calendar/CalendarMiniMonthDay.vue
